Move authentication, wrap and pagination to middleware folder
